♻️ Refactor: Move playcount feature to MVC pattern
